fix: overhaul zombie games - fix controls, organization, and damage registration
**Controls Fixed (All Games):**
- Swapped mouse controls to be more intuitive:
  - Right-click: Aim/draw weapon (narrows FOV)
  - Left-click: Fire weapon
- Updated all in-game instructions to reflect new controls

**Damage Registration Improvements:**
- Increased aim tolerance in doom-raycaster for better hit detection
  - Not aiming: 0.2 → 0.25 radians
  - Aiming: 0.1 → 0.15 radians
  - Distance bonus: 0.05 → 0.08
- This makes combat more responsive and reduces frustration

**Major Reorganization:**
- Created `games/zombie-shooters/` directory with properly named games:
  - `doom-raycaster/` - Classic raycasting Doom-style FPS
  - `hallway-survival/` - 3D corridor shooter with dual weapons
  - `hallway-enhanced/` - Enhanced 3D shooter with fireballs
- Archived old/duplicate versions to `games/archived-versions/`:
  - CDN-based versions (potential black screen issues)
  - Multiple fix attempts from previous PRs
- Updated all game titles and descriptions for clarity

**Documentation:**
- Added comprehensive README.md to zombie-shooters folder
- Documents all three games, controls, and features
- Includes technical notes about local vendor files

**Files Changed:**
- All game controls swapped (mousedown/mouseup events)
- All HTML updated with correct control instructions
- All game titles updated to descriptive names
- Directory structure completely reorganized
- Messy duplicate files archived

This resolves the issues from PRs #13-16 by consolidating fixes
and creating a clean, maintainable structure.
